jhiRoot/plantsFront/src/main/webapp/app/entities/plantsMS/clade/route/clade-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IClade, Clade } from '../clade.model';
import { CladeService } from '../service/clade.service';

import { CladeRoutingResolveService } from './clade-routing-resolve.service';

describe('Clade routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: CladeRoutingResolveService;
  let service: CladeService;
  let resultClade: IClade | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(CladeRoutingResolveService);
    service = TestBed.inject(CladeService);
    resultClade = undefined;
  });

  describe('resolve', () => {
    it('should return IClade returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultClade = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultClade).toEqual({ id: 123 });
    });

    it('should return new IClade if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultClade = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultClade).toEqual(new Clade());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Clade })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultClade = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultClade).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
